// src/app/models/iPersonas.ts
export interface iPersona{
  _id?: string;
  email?: string;
  apellido?: string;
  nombre?: string;
  name?: string;
  cuit?: string;
  fijo?: string;
  celular?: string;
  direccion?: string;
  localidad?: string;
  provincia?: string;
  zipcode?: string;
  pais?: string;
  coeficiente?: number;
  categoria?: [];
}